+++ steampipecloud/table_steampipecloud_workspace_mod.go
@@ -53,7 +53,7 @@ func tableSteampipeCloudWorkspaceMod(_ context.Context) *plugin.Table {
 				Name:        "mod_constraint",
 				Description: "Version constraint for the mod.",
 				Type:        proto.ColumnType_STRING,
-				Transform:   transform.FromCamel(),
+				Transform:   transform.FromField("Constraint"),
 			},
 			{
 				Name:        "alias",
@@ -102,7 +102,7 @@ func tableSteampipeCloudWorkspaceMod(_ context.Context) *plugin.Table {
 //// LIST FUNCTION
 
 func listWorkspaceMods(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-	workspace := h.Item.(openapi.Workspace)
+	workspace := h.Item.(*openapi.Workspace)
 
 	// Create Session
 	svc, err := connect(ctx, d)
@@ -158,12 +158,12 @@ func listUserWorkspaceMods(ctx context.Context, d *plugin.QueryData, h *plugin.H
 	for pagesLeft {
 		if resp.NextToken != nil {
 			listDetails = func(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-				resp, _, err = svc.UserWorkspaceMods.List(context.Background(), userHandle, workspaceHandle).NextToken(*resp.NextToken).Limit(maxResults).Execute()
+				resp, _, err = svc.UserWorkspaceMods.List(ctx, userHandle, workspaceHandle).NextToken(*resp.NextToken).Limit(maxResults).Execute()
 				return resp, err
 			}
 		} else {
 			listDetails = func(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-				resp, _, err = svc.UserWorkspaceMods.List(context.Background(), userHandle, workspaceHandle).Limit(maxResults).Execute()
+				resp, _, err = svc.UserWorkspaceMods.List(ctx, userHandle, workspaceHandle).Limit(maxResults).Execute()
 				return resp, err
 			}
 		}
@@ -181,9 +181,6 @@ func listUserWorkspaceMods(ctx context.Context, d *plugin.QueryData, h *plugin.H
 			for _, workspaceMod := range *result.Items {
 				workspaceMod.Workspace = &openapi.Workspace{}
 				workspaceMod.Workspace.Handle = workspaceHandle
-				workspaceMod.Workspace.Identity = &openapi.Identity{}
-				workspaceMod.Workspace.Identity.Handle = userHandle
-				workspaceMod.Workspace.Identity.Type = "user"
 				d.StreamListItem(ctx, workspaceMod)
 
 				// Context can be cancelled due to manual cancellation or the limit has been hit
@@ -213,12 +210,12 @@ func listOrgWorkspaceMods(ctx context.Context, d *plugin.QueryData, h *plugin.Hy
 	for pagesLeft {
 		if resp.NextToken != nil {
 			listDetails = func(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-				resp, _, err = svc.OrgWorkspaceMods.List(context.Background(), orgHandle, workspaceHandle).NextToken(*resp.NextToken).Limit(maxResults).Execute()
+				resp, _, err = svc.OrgWorkspaceMods.List(ctx, orgHandle, workspaceHandle).NextToken(*resp.NextToken).Limit(maxResults).Execute()
 				return resp, err
 			}
 		} else {
 			listDetails = func(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-				resp, _, err = svc.OrgWorkspaceMods.List(context.Background(), orgHandle, workspaceHandle).Limit(maxResults).Execute()
+				resp, _, err = svc.OrgWorkspaceMods.List(ctx, orgHandle, workspaceHandle).Limit(maxResults).Execute()
 				return resp, err
 			}
 		}
@@ -236,9 +233,6 @@ func listOrgWorkspaceMods(ctx context.Context, d *plugin.QueryData, h *plugin.Hy
 			for _, workspaceMod := range *result.Items {
 				workspaceMod.Workspace = &openapi.Workspace{}
 				workspaceMod.Workspace.Handle = workspaceHandle
-				workspaceMod.Workspace.Identity = &openapi.Identity{}
-				workspaceMod.Workspace.Identity.Handle = orgHandle
-				workspaceMod.Workspace.Identity.Type = "org"
 				d.StreamListItem(ctx, workspaceMod)
 
 				// Context can be cancelled due to manual cancellation or the limit has been hit
@@ -311,7 +305,7 @@ func getUserWorkspaceMod(ctx context.Context, d *plugin.QueryData, h *plugin.Hyd
 	var resp openapi.WorkspaceMod
 
 	getDetails := func(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-		resp, _, err = svc.UserWorkspaceMods.Get(context.Background(), identityId, workspaceId, alias).Execute()
+		resp, _, err = svc.UserWorkspaceMods.Get(ctx, identityId, workspaceId, alias).Execute()
 		return resp, err
 	}
 
@@ -319,8 +313,6 @@ func getUserWorkspaceMod(ctx context.Context, d *plugin.QueryData, h *plugin.Hyd
 
 	workspaceMod := response.(openapi.WorkspaceMod)
 	workspaceMod.Workspace = &openapi.Workspace{}
-	workspaceMod.Workspace.Identity = &openapi.Identity{}
-	workspaceMod.Workspace.Identity.Type = "user"
 
 	if err != nil {
 		plugin.Logger(ctx).Error("getUserWorkspaceMod", "get", err)
@@ -337,7 +329,7 @@ func getOrgWorkspaceMod(ctx context.Context, d *plugin.QueryData, h *plugin.Hydr
 	var resp openapi.WorkspaceMod
 
 	getDetails := func(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-		resp, _, err = svc.OrgWorkspaceMods.Get(context.Background(), identityId, workspaceId, alias).Execute()
+		resp, _, err = svc.OrgWorkspaceMods.Get(ctx, identityId, workspaceId, alias).Execute()
 		return resp, err
 	}
 
@@ -345,8 +337,6 @@ func getOrgWorkspaceMod(ctx context.Context, d *plugin.QueryData, h *plugin.Hydr
 
 	workspaceMod := response.(openapi.WorkspaceMod)
 	workspaceMod.Workspace = &openapi.Workspace{}
-	workspaceMod.Workspace.Identity = &openapi.Identity{}
-	workspaceMod.Workspace.Identity.Type = "org"
 
 	if err != nil {
 		plugin.Logger(ctx).Error("getOrgWorkspaceMod", "get", err)