+++ steampipecloud/table_steampipecloud_workspace_mod_variable.go
@@ -130,9 +130,7 @@ func listWorkspaceModVariables(ctx context.Context, d *plugin.QueryData, h *plug
 		return nil, nil
 	}
 
-	workspace := h.Item.(openapi.Workspace)
-
-	plugin.Logger(ctx).Trace("listWorkspaceModVariables", "workspace.Id", workspace.Id)
+	workspace := h.Item.(*openapi.Workspace)
 
 	if workspace.Id != workspaceId {
 		return nil, nil
@@ -192,12 +190,12 @@ func listUserWorkspaceModVariables(ctx context.Context, d *plugin.QueryData, h *
 	for pagesLeft {
 		if resp.NextToken != nil {
 			listDetails = func(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-				resp, _, err = svc.UserWorkspaceModVariables.List(context.Background(), userHandle, workspaceHandle, modAlias).NextToken(*resp.NextToken).Limit(maxResults).Execute()
+				resp, _, err = svc.UserWorkspaceModVariables.List(ctx, userHandle, workspaceHandle, modAlias).NextToken(*resp.NextToken).Limit(maxResults).Execute()
 				return resp, err
 			}
 		} else {
 			listDetails = func(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-				resp, _, err = svc.UserWorkspaceModVariables.List(context.Background(), userHandle, workspaceHandle, modAlias).Limit(maxResults).Execute()
+				resp, _, err = svc.UserWorkspaceModVariables.List(ctx, userHandle, workspaceHandle, modAlias).Limit(maxResults).Execute()
 				return resp, err
 			}
 		}
@@ -242,12 +240,12 @@ func listOrgWorkspaceModVariables(ctx context.Context, d *plugin.QueryData, h *p
 	for pagesLeft {
 		if resp.NextToken != nil {
 			listDetails = func(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-				resp, _, err = svc.OrgWorkspaceModVariables.List(context.Background(), orgHandle, workspaceHandle, modAlias).NextToken(*resp.NextToken).Limit(maxResults).Execute()
+				resp, _, err = svc.OrgWorkspaceModVariables.List(ctx, orgHandle, workspaceHandle, modAlias).NextToken(*resp.NextToken).Limit(maxResults).Execute()
 				return resp, err
 			}
 		} else {
 			listDetails = func(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-				resp, _, err = svc.OrgWorkspaceModVariables.List(context.Background(), orgHandle, workspaceHandle, modAlias).Limit(maxResults).Execute()
+				resp, _, err = svc.OrgWorkspaceModVariables.List(ctx, orgHandle, workspaceHandle, modAlias).Limit(maxResults).Execute()
 				return resp, err
 			}
 		}